Check whether Boost.Graph supports the current compiler.
Recent change to Boost.Math, on which Boost.Graph depends, made it
incompatible with MSVC 14.0. Rather than check for a specific compiler
version, add a configure-time check whether Boost.Graph compiles (or
at least the part that is used in Boost.Parameter tests). Disable the
tests that depend on Boost.Graph if that check fails.

Refs https://github.com/boostorg/math/issues/1275.